backend/requirements.txt
Flask
requests
python-dotenv
cachetools
flask-cors
